Moved UT to injectSheet.test.js
