+++ src/injectSheet.test.js
@@ -6,6 +6,8 @@ import {create} from 'jss'
 import getDisplayName from './getDisplayName'
 import createHoc from './createHoc'
 import {createGenerateClassName} from '../tests/helper'
+import {stripIndent} from 'common-tags'
+import preset from 'jss-preset-default'
 
 describe('injectSheet', () => {
   let jss
@@ -276,6 +278,76 @@ describe('injectSheet', () => {
       render(<MyComponent classes={{user: 'user'}} />, node)
       expect(classes).to.eql({default: 'default', a: 'a-id', user: 'user'})
     })
+
+    it('setState should use new context', () => {
+      const ThemeA = {color: "blue"};
+      const ThemeB = {color: "red"};
+      const ComponentA = injectSheet(() => ({a: {left: 2}}))()
+      const localJss = createJss({
+        ...preset(),
+        createGenerateClassName: () => {
+          let counter = 0
+          return rule => `${rule.key}-${counter++}`
+        }
+      })
+      render((
+        <JssProvider jss={localJss}>
+          <ThemeProvider theme={ThemeA}>
+            <div>
+              <ComponentA />
+            </div>
+          </ThemeProvider>
+        </JssProvider>
+      ), node)
+  
+      const styleTags = Array.from(document.querySelectorAll('style'))
+      const innerText = x => x.innerText
+      const trim = x => x.trim()
+      const actual = styleTags.map(innerText).map(trim).join('\n')
+  
+      expect(actual).to.be(stripIndent`
+        .a-0 {
+          left: 2px;
+        }
+      `)
+      function rtl() {
+        return {
+          onProcessStyle(style, rule, sheet) {
+            return {
+              right: '2px'
+            }
+          }
+        }
+      }
+      const newJss = createJss({
+        ...preset(),
+        createGenerateClassName: () => {
+          let counter = 0
+          return rule => `${rule.key}-${counter++}`
+        }
+      })
+      newJss.use(rtl())
+      render((
+        <JssProvider jss={newJss}>
+          <ThemeProvider theme={ThemeB}>
+            <div>
+              <ComponentA />
+            </div>
+          </ThemeProvider>
+        </JssProvider>
+      ), node)
+      const newStyleTags = Array.from(document.querySelectorAll('style'))
+      const newInnerText = x => x.innerText
+      const newTrim = x => x.trim()
+      const newActual = newStyleTags.map(newInnerText).map(newTrim).join('\n')
+  
+      expect(newActual).to.be(stripIndent`
+        .a-1 {
+          right: 2px;
+        }
+      `)
+    })
+  
   })
 
   describe('classNamePrefix', () => {
